add support for acknowledge and resolve
Support for resolving and acknowledging incidents has been added.
Although, the latter is probably not all that useful. When using the
trigger, acknowledge, and resolve function calls the gen_server will
make attempts to retry calls to the pagerduty API when appropriate.

Added meck to aid testing and upgraded the jsx dependency.